fix: fixed "x is a custom class, make sure it is defined within a `@layer` directive" error
I think there were several issues that were causing this:
- The importing project was importing from assets/*.css which points to dist
instead of src like /*css.
- Added imports for style.css.
- I don't think plugins were getting correctly added when using the nuxt hook.
Switched to the suggested method of using a template. Only issue is
metamorphosis won't work this way (due to some jiti stuff with an esm only dep I
think) so added it's theme config manually from the module. This should also
work better if the user specifies a config.

Added divs in the playground app to test this is all working correctly.